Improve group select

+++ avr/cores/megacommand/MCL/GridIOPage.h
@@ -7,7 +7,7 @@
 
 class GridIOPage : public LightPage {
  public:
- static uint8_t track_type_select;
+ static uint16_t track_type_select;
  static bool show_track_type_select;
 
  GridIOPage(Encoder *e1 = NULL, Encoder *e2 = NULL, Encoder *e3 = NULL, Encoder *e4 = NULL) : LightPage( e1, e2, e3 ,e4) {
